Add missing assignment operator to `...precision`
Some compilers will warn about an inability to construct a default
assignment operator for the `munit_plus_precision` class, due to
the reference-type data-member. This commit adds a "deleted"
assignment operator to the class, to communicate the intent of not
providing any assignment operator for this class.

Addresses codylico/munit-plus#16.