Add extra exception check to `...error_jmp`
Visual Studio 2010 seems to not have any way of inspecting an
`std::exception_ptr` to see if it is empty. Therefore, considering
the perceived-as-documented behavior of throwing a
`std::bad_exception` for empty `std::exception_ptr` instances,
and weighing the necessity of avoiding ignoring excpetions in
general, this commit hard-codes the `munit_plus_error_jmp` to
always report a nested exception.

This problem should be revisited when someone discovers a way to
detect active exceptions.

Addresses issue codylico/munit-plus#15.